+++ client/state/login/actions/index.js
@@ -13,15 +13,10 @@ import {
 	LOGIN_AUTH_ACCOUNT_TYPE_REQUEST_FAILURE,
 	LOGIN_AUTH_ACCOUNT_TYPE_RESET,
 	LOGIN_FORM_UPDATE,
-	TWO_FACTOR_AUTHENTICATION_PUSH_POLL_START,
-	TWO_FACTOR_AUTHENTICATION_PUSH_POLL_STOP,
-	TWO_FACTOR_AUTHENTICATION_PUSH_POLL_COMPLETED,
 } from 'state/action-types';
 import { recordTracksEventWithClientId as recordTracksEvent } from 'state/analytics/actions';
 import 'state/data-layer/wpcom/login-2fa';
 import 'state/data-layer/wpcom/users/auth-options';
-
-import { remoteLoginUser } from 'state/login/actions/remote-login-user';
 
 import 'state/login/init';
 
@@ -35,9 +30,12 @@ export { connectSocialUser } from 'state/login/actions/connect-social-user';
 export { disconnectSocialUser } from 'state/login/actions/disconnect-social-user';
 export { createSocialUserFailed } from 'state/login/actions/create-social-user-failed';
 export { sendSmsCode } from 'state/login/actions/send-sms-code';
+export {
+	receivedTwoFactorPushNotificationApproved,
+	startPollAppPushAuth,
+	stopPollAppPushAuth,
+} from 'state/login/actions/push';
 
-export const startPollAppPushAuth = () => ( { type: TWO_FACTOR_AUTHENTICATION_PUSH_POLL_START } );
-export const stopPollAppPushAuth = () => ( { type: TWO_FACTOR_AUTHENTICATION_PUSH_POLL_STOP } );
 export const formUpdate = () => ( { type: LOGIN_FORM_UPDATE } );
 
 /**
@@ -92,19 +90,3 @@ export const getAuthAccountType = usernameOrEmail => dispatch => {
 export const resetAuthAccountType = () => ( {
 	type: LOGIN_AUTH_ACCOUNT_TYPE_RESET,
 } );
-
-/**
- * Creates an action that indicates that the push poll is completed
- *
- * @param {Array<string>} tokenLinks token links array
- * @returns {Function} a thunk
- */
-export const receivedTwoFactorPushNotificationApproved = tokenLinks => dispatch => {
-	if ( ! Array.isArray( tokenLinks ) ) {
-		return dispatch( { type: TWO_FACTOR_AUTHENTICATION_PUSH_POLL_COMPLETED } );
-	}
-
-	remoteLoginUser( tokenLinks ).then( () =>
-		dispatch( { type: TWO_FACTOR_AUTHENTICATION_PUSH_POLL_COMPLETED } )
-	);
-};
